rse: css-aspen: Add SCMI Comms secure partition
This patch enables the use of SCMI Comms for
RSE runtime. It expands upon the SCMI HAL library
to include functions for handling the shared
memory. The shared memory is partitioned into
two pages. One for sending and receiving commands
from a SCP (a.k.a the platform), and receiving
notification from SCP.

Implementing the protocol in this way allows the
agent to receive notifications pertaining to shutdown
and reset, in accordance with SCP SCMI best practice.
Both pages of shared memory are initialised and are
visible to the HAL library exclusively.

This patch depends upon the SCMI Comms being available
for BL2 and the TF-M Extras source code of SCMI. During
the build, the configuration will differentiate
between what is required for BL2 and RSE runtime.
Functions for writing to shared memory are written but
not used due to issues with its functionality.

Upstream-Status: Pending

Signed-off-by: Mubashar Ahmad <mubashar.ahmad@arm.com>
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>