rse: Route IC trigger to trigger output interface
DMA ICS program 4 command 4 hangs indefinitely waiting for
trigger 5 from the integrity checker, this happens because
TRIG_MUX_CTRL of the ILCU is not configured to route the
integrity checker trigger to the trigger output interface.

Add an extra DMA ICS command before triggering the integrity
checker to do the above.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Signed-off-by: Ambroise Vincent <ambroise.vincent@arm.com>