rse: css-aspen: Enable SI UART debug allowed bit
In the Safety Island, the UARTs are not protected by any safety
mechanism, as they are not safety relevant. To ensure there is no
interference from the UARTs in the case of an error (e.g. a spurious
interrupt to the GIC), the UART outputs are safely gated with the
DEBUG_ALLOWED signal from the SCR.

Enable DEBUG_ALLOWED signal to see UART messages.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>