plat:css-aspen: Workaround ESRT construction
ESRT is populated and installed for Linux before new images are accepted
in ExitBootServices(). The current implementation updates ESRT
information only after new images are accepted or rejected. Therefore,
ESRT shows out-of-date information when Linux reads it. Out-of-date ESRT
also breaks SystemReady ACS Capsule Update test suite.

As a workaround, mark the last attempt status as SUCCESS when the
component images are installed. Therefore, TF-M can return SUCCESS
status in ESRT reported to test suites or Linux.
If the images are rejected or fail to boot, update the last attempt
statuses to failures.
Also initialize LowestSupportedFwVersion during ESRT construction if it
is not provisioned yet.

Upstream-Status: Pending

Signed-off-by: David Hu <david.hu2@arm.com>